separate files out in prep for registry
