+++ components/supabase-manager/logs.tsx
@@ -29,7 +29,7 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
-import { useGetLogs } from "@/hooks/use-supabase-manager";
+import { useGetLogs } from "@/hooks/use-logs";
 import { LogsTableName, genDefaultQuery } from "@/lib/logs";
 import { cn } from "@/lib/utils";
 import { Check, ChevronsUpDown, Logs, Terminal } from "lucide-react";
