+++ components/supabase-manager/users-growth.tsx
@@ -8,7 +8,7 @@ import {
   ChartTooltip,
   ChartTooltipContent,
 } from "@/components/ui/chart";
-import { useGetUserCountsByDay } from "@/hooks/use-supabase-manager";
+import { useGetUserCountsByDay } from "@/hooks/use-user-counts";
 import { Skeleton } from "@/components/ui/skeleton";
 import { AlertTriangle } from "lucide-react";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
