+++ hooks/use-run-query.ts
@@ -1,0 +1,51 @@
+"use client";
+
+import { client } from "@/lib/management-api";
+import type { components } from "@/lib/management-api-schema";
+import { listTablesSql } from "@/lib/pg-meta";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { type AxiosError } from "axios";
+import { toast } from "sonner";
+
+// RUN SQL Query
+export const runQuery = async ({
+  projectRef,
+  query,
+  readOnly,
+}: {
+  projectRef: string;
+  query: string;
+  readOnly?: boolean;
+}) => {
+  const { data, error } = await client.POST(
+    "/v1/projects/{ref}/database/query",
+    {
+      params: {
+        path: {
+          ref: projectRef,
+        },
+      },
+      body: {
+        query,
+        read_only: readOnly,
+      },
+    }
+  );
+
+  if (error) {
+    throw error;
+  }
+
+  return data as any;
+};
+
+export const useRunQuery = () => {
+  return useMutation({
+    mutationFn: runQuery,
+    onError: (error: AxiosError<{ message: string }>) => {
+      toast.error(
+        error.response?.data?.message || "There was a problem with your query."
+      );
+    },
+  });
+};
