+++ hooks/use-suggestions.ts
@@ -1,0 +1,53 @@
+"use client";
+
+import { client } from "@/lib/management-api";
+import { useQuery } from "@tanstack/react-query";
+
+// GET Suggestions
+const getSuggestions = async (projectRef: string) => {
+  const [
+    { data: performanceData, error: performanceError },
+    { data: securityData, error: securityError },
+  ] = await Promise.all([
+    client.GET("/v1/projects/{ref}/advisors/performance", {
+      params: {
+        path: {
+          ref: projectRef,
+        },
+      },
+    }),
+    client.GET("/v1/projects/{ref}/advisors/security", {
+      params: {
+        path: {
+          ref: projectRef,
+        },
+      },
+    }),
+  ]);
+  if (performanceError) {
+    throw performanceError;
+  }
+  if (securityError) {
+    throw securityError;
+  }
+
+  // Add type to each suggestion
+  const performanceLints = (performanceData?.lints || []).map((lint) => ({
+    ...lint,
+    type: "performance" as const,
+  }));
+  const securityLints = (securityData?.lints || []).map((lint) => ({
+    ...lint,
+    type: "security" as const,
+  }));
+  return [...performanceLints, ...securityLints];
+};
+
+export const useGetSuggestions = (projectRef: string) => {
+  return useQuery({
+    queryKey: ["suggestions", projectRef],
+    queryFn: () => getSuggestions(projectRef),
+    enabled: !!projectRef,
+    retry: false,
+  });
+};
